+++ cli/internal/helm/generateCilium.sh
@@ -3,7 +3,7 @@
 CALLDIR=$(pwd)
 CILIUMTMPDIR=$(mktemp -d)
 cd $CILIUMTMPDIR
-git clone --depth 1 -b v1.12 https://github.com/cilium/cilium.git
+git clone --depth 1 -b 1.12.1 https://github.com/cilium/cilium.git
 cd cilium
 git apply $CALLDIR/cilium.patch
 cp -r install/kubernetes/cilium $CALLDIR/charts