+++ cli/internal/helm/values.go
@@ -16,27 +16,28 @@ var azureVals = map[string]interface{}{
 			},
 		},
 	},
-	"strictModeCIDRs": []string{
-		"10.244.0.0/16",
-	},
+	"strictModeCIDR": "10.244.0.0/16",
 	"image": map[string]interface{}{
 		"repository": "ghcr.io/3u13r/cilium",
-		"suffix":     "v1.12.0-edg2",
-		"tag":        "latest",
-		"digest":     "sha256:8dee8839bdf4cfdc28a61c4586f23f2dbfabe03f94dee787c4d749cfcc02c6bf",
-		"useDigest":  false,
+		"suffix":     "",
+		"tag":        "v1.12.1-edg",
+		"digest":     "sha256:fdac430143fe719331698b76fbe66410631a21afd3405407d56db260d2d6999b",
+		"useDigest":  true,
 	},
 	"operator": map[string]interface{}{
 		"image": map[string]interface{}{
 			"repository":    "ghcr.io/3u13r/operator",
-			"tag":           "v1.12.0-edg2",
+			"tag":           "v1.12.1-edg",
 			"suffix":        "",
-			"genericDigest": "sha256:adbdeb0199aa1d870940c3363bfa5b69a5c8b4f533fc9f67463f8d447077464a",
+			"genericDigest": "sha256:a225d8d3976fd2a05cfa0c929cd32e60283abedf6bae51db4709df19b2fb70cb",
 			"useDigest":     true,
 		},
 	},
-	"egressMasqueradeInterfaces": "eth0",
-	"enableIPv4Masquerade":       true,
+	"egressMasqueradeInterfaces":          "eth0",
+	"enableIPv4Masquerade":                true,
+	"kubeProxyReplacement":                "strict",
+	"enableCiliumEndpointSlice":           true,
+	"kubeProxyReplacementHealthzBindAddr": "0.0.0.0:10256",
 }
 
 var gcpVals = map[string]interface{}{
@@ -51,16 +52,16 @@ var gcpVals = map[string]interface{}{
 	"image": map[string]interface{}{
 		"repository": "ghcr.io/3u13r/cilium",
 		"suffix":     "",
-		"tag":        "v1.12.0-edg2",
-		"digest":     "sha256:8dee8839bdf4cfdc28a61c4586f23f2dbfabe03f94dee787c4d749cfcc02c6bf",
+		"tag":        "v1.12.1-edg",
+		"digest":     "sha256:fdac430143fe719331698b76fbe66410631a21afd3405407d56db260d2d6999b",
 		"useDigest":  true,
 	},
 	"operator": map[string]interface{}{
 		"image": map[string]interface{}{
 			"repository":    "ghcr.io/3u13r/operator",
 			"suffix":        "",
-			"tag":           "v1.12.0-edg2",
-			"genericDigest": "sha256:adbdeb0199aa1d870940c3363bfa5b69a5c8b4f533fc9f67463f8d447077464a",
+			"tag":           "v1.12.1-edg",
+			"genericDigest": "sha256:a225d8d3976fd2a05cfa0c929cd32e60283abedf6bae51db4709df19b2fb70cb",
 			"useDigest":     true,
 		},
 	},
@@ -68,6 +69,9 @@ var gcpVals = map[string]interface{}{
 	"ipam": map[string]interface{}{
 		"mode": "kubernetes",
 	},
+	"kubeProxyReplacement":                "strict",
+	"enableCiliumEndpointSlice":           true,
+	"kubeProxyReplacementHealthzBindAddr": "0.0.0.0:10256",
 }
 
 var qemuVals = map[string]interface{}{